requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
sqlcipher3-wheels==0.5.5.post0